Move python code to seperate file
